% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwrig.r
\docType{package}
\name{bigwig-package}
\alias{bigwig}
\alias{bigwig-package}
\title{bigwig: read data from bigWig files}
\description{
bigwig provides methods to read data from bigWig files. bigwig uses
Rcpp to wrap libBigWig from @dpryan79.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://rnabioco.github.io/bigwig/}
}

}
\author{
Jay Hesselberth <jay.hesselberth@gmail.com>
}
