% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwrig.r
\docType{package}
\name{bigwrig}
\alias{bigwrig}
\alias{bigwrig-package}
\title{bigwrig: read data from bigWig files}
\description{
bigwrig provides methods to read data from bigWig files. bigWrig uses
Rcpp to wrap libBigWig from @dpryan79.
}
\author{
Jay Hesselberth <jay.hesselberth@gmail.com>
}

