% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig-package.R
\docType{package}
\name{cpp11bigwig-package}
\alias{cpp11bigwig}
\alias{cpp11bigwig-package}
\title{cpp11bigwig: read data from bigWig files}
\description{
bigwig provides methods to read data from bigWig files. bigwig uses
cpp11 to wrap libBigWig from @dpryan79.
}
\details{
<https://github.com/dpryan79/libBigWig>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/cpp11bigwig/}
  \item \url{https://github.com/rnabioco/cpp11bigwig}
}

}
\author{
Jay Hesselberth <jay.hesselberth@gmail.com>
}
\keyword{internal}
