% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwrig.r
\name{read_bigwig}
\alias{read_bigwig}
\title{Read data from bigWig files.}
\usage{
read_bigwig()
}
\arguments{
\item{bwfile}{filename or URL for bigWig file}

\item{chrom}{read data for specific chromsome}
}
\value{
\code{data_frame}
}
\description{
Read data from bigWig files.
}
\examples{

}
\seealso{
\url{https://github.com/dpryan79/libBigWig}

\url{https://github.com/brentp/bw-python}
}

