% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bigwig.r
\name{read_bigwig}
\alias{read_bigwig}
\title{Read data from bigWig files.}
\usage{
read_bigwig(bwfile, chrom = NULL, start = NULL, end = NULL, raw = FALSE)
}
\arguments{
\item{bwfile}{filename for bigWig file}

\item{chrom}{read data for specific chromosome}

\item{start}{start position for data}

\item{end}{end position for data}

\item{raw}{return raw data, single-base intervals}
}
\value{
\code{data.frame}
}
\description{
Read data from bigWig files.
}
\examples{
bw <- system.file('extdata', 'test.bw', package = 'cpp11bigwig')
read_bigwig(bw)

}
\seealso{
\url{https://github.com/dpryan79/libBigWig}

\url{https://github.com/brentp/bw-python}
}
